requirements.txt
streamlit
yt-dlp
openai-whisper
demucs
Pillow
opencv-python-headless
numpy
requests
watchdog
